requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4  
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0   
APScheduler==3.11.0      
asyncio==3.4.3
attrs==25.1.0
beautifulsoup4==4.12.3   
blinker==1.9.0
bs4==0.0.2
certifi==2024.12.14      
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.0
Flask-Cors==5.0.0
frozenlist==1.5.0
humanize==4.11.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
motor==3.6.1
multidict==6.1.0
packaging==24.2
pluggy==1.5.0
propcache==0.2.1
pyaes==1.6.1
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.9.2
Pyrogram==2.0.106
pyromod==3.1.6
PySocks==1.7.1
pytest==8.3.4
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
schedule==1.2.2
shortzy==0.0.8
soupsieve==2.6
TgCrypto==1.2.5
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.18.3